Add splash screen for initial token check
