+++ app/src/main/java/com/cicero/repostapp/MainActivity.kt
@@ -4,12 +4,6 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import android.widget.Button
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-import okhttp3.OkHttpClient
-import okhttp3.Request
 
 class MainActivity : AppCompatActivity() {
 
@@ -21,46 +15,10 @@ class MainActivity : AppCompatActivity() {
         supportActionBar?.setDisplayUseLogoEnabled(true)
 
 
-        val prefs = getSharedPreferences("auth", MODE_PRIVATE)
-        val token = prefs.getString("token", null)
-        val userId = prefs.getString("userId", null)
-        if (!token.isNullOrBlank() && !userId.isNullOrBlank()) {
-            validateToken(token, userId)
-        }
 
         findViewById<Button>(R.id.button_open_login).setOnClickListener {
             startActivity(Intent(this, LoginActivity::class.java))
         }
-    }
-
-    private fun validateToken(token: String, userId: String) {
-        CoroutineScope(Dispatchers.IO).launch {
-            val client = OkHttpClient()
-            val request = Request.Builder()
-                .url("https://papiqo.com/api/users/$userId")
-                .header("Authorization", "Bearer $token")
-                .build()
-            try {
-                client.newCall(request).execute().use { resp ->
-                    withContext(Dispatchers.Main) {
-                        if (resp.isSuccessful) {
-                            navigateToDashboard(token, userId)
-                        }
-                    }
-                }
-            } catch (_: Exception) {
-                // ignore
-            }
-        }
-    }
-
-    private fun navigateToDashboard(token: String, userId: String) {
-        val intent = Intent(this, DashboardActivity::class.java).apply {
-            putExtra("token", token)
-            putExtra("userId", userId)
-        }
-        startActivity(intent)
-        finish()
     }
 
     // No external storage permission required when using app-specific storage
